Moved linearity subfolder into queries folder
