+++ vehicle/src/Vehicle/Compile/Queries/Linearity.hs
@@ -1,4 +1,4 @@
-module Vehicle.Compile.Linearity
+module Vehicle.Compile.Queries.Linearity
   ( module X
   , solveForUserVariables
   , reconstructUserVars
@@ -12,10 +12,10 @@ import Data.Vector.Unboxed (Vector)
 import Data.Vector.Unboxed qualified as Vector
 
 import Vehicle.Compile.Error
-import Vehicle.Compile.Linearity.Core as X
-import Vehicle.Compile.Linearity.FourierMotzkinElimination (fourierMotzkinElimination,
+import Vehicle.Compile.Queries.Linearity.Core as X
+import Vehicle.Compile.Queries.Linearity.FourierMotzkinElimination (fourierMotzkinElimination,
                                                             reconstructFMUserVar)
-import Vehicle.Compile.Linearity.GaussianElimination (gaussianElimination)
+import Vehicle.Compile.Queries.Linearity.GaussianElimination (gaussianElimination)
 import Vehicle.Compile.Prelude
 
 solveForUserVariables :: MonadCompile m
